+++ packages/widgetbook/lib/src/knobs/knobs.dart
@@ -13,6 +13,6 @@ export 'int_input_knob.dart';
 export 'int_slider_knob.dart';
 export 'knob.dart';
 export 'knobs_registry.dart';
-export 'list_knob.dart';
+export 'object_dropdown_knob.dart';
 export 'object_segmented_knob.dart';
 export 'string_knob.dart';